requirements.txt
flask
python-dotenv
sqlalchemy
mysqlclient
pycurl
pandas
requests
pysftp